update: 대시보드에 필요한 정보 조회 기능 추가

+++ src/main/java/org/example/commercebackoffice/dashboard/controller/dto/response/ReviewRating.java
@@ -1,7 +1,7 @@
 package org.example.commercebackoffice.dashboard.controller.dto.response;
 
 public record ReviewRating(
-        int rating,
-        int count
+        Integer rating,
+        Long count
 ) {
 }
